part4.14-4.23

+++ part4/controllers/blogs.js
@@ -1,9 +1,12 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 // get all
 blogsRouter.get('/', async (req, res) => {
   const blogs = await Blog.find({})
+    .populate('user', { username: 1, name: 1 })
   res.json(blogs.map(blog => blog.toJSON()))
 })
 
@@ -19,21 +22,44 @@ blogsRouter.get('/:id', async (req, res) => {
 // post
 blogsRouter.post('/', async (req, res) => {
   const body = req.body
+  //const token = tokenExtractor(req)
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return res.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+  const user = req.user
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: user._id,
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   res.json(savedBlog.toJSON())
 })
 
 // delete
 blogsRouter.delete('/:id', async (req, res) => {
-  await Blog.findByIdAndRemove(req.params.id)
-  res.status(204).end()
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  const user = req.user
+  const blog = await Blog.findById(req.params.id)
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(req.params.id)
+    res.status(204).end()
+  } 
+  if (!decodedToken.id) {
+    return res.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
 })
 
 // update
